+++ socket.h
@@ -34,12 +34,12 @@
 #include "interface.h"
 
 void make_dhcp_packet(struct udp_dhcp_packet *packet,
-		      const unsigned char *data, int length,
+		      const unsigned char *data, size_t length,
 		      struct in_addr source, struct in_addr dest);
 
 int open_socket (interface_t *iface, bool arp);
 ssize_t send_packet (const interface_t *iface, int type,
-		     const unsigned char *data, int len);
+		     const unsigned char *data, size_t len);
 ssize_t get_packet (const interface_t *iface, unsigned char *data,
-		    unsigned char *buffer, int *buffer_len, int *buffer_pos);
+		    unsigned char *buffer, size_t *buffer_len, size_t *buffer_pos);
 #endif
